Move the data for the resume page to a RON file
When a request is made to `/resume` the server will read the file and
parse it to generate the `ResumePage` template. This allows for the file
to be changed without requiring a server restart.

When an 500 is returned using the `AppError` it now uses debug print
rather than display since this will show downcast errors.